Big commit
Okay, so in this commit I did a bunch of things. Probably should have
done multiple, small commits but... oh well.

First, I redesigned the linked list data structure from the first Cards
attempt (deck.h/c). One of the problems I was running into in the first
program was that a deck was just a linked list of cards... But if there
was no cards in the deck the deck couldn't exist. So, in this program, a
Deck is a structure that contains (manages?) a linked list of cards.
This structure can than be used for the player's hand, tableaus on the
table, foundation piles... whatever. The structure has an integer id
that will (eventually) correspond to buttons on the screen, and a count
that is kept up to date when cards are added/removed. There is, of
course, a function to count all the cards by going through each one...
but this (might be) is convienent.

Additionally, I added a couple functions for printing out the cards -
one that prints just the suite and rank in color, and one that prints
out a 4x4 card with a white background.

I also copied over the mt19937 files from the Toolbox, and the flags file
from the previous Cards project.