Global settings, buttons
I was getting irritated with changing the color of the cards - realized
it would be SUPER simple to just throw all the settings in a global
struct. I suppose eventually these could even be set "in game" - like on
an options screen.

Added in simple "buttons", that can be toggled selected and
activated/deactivated.