Updated Klondike main loop
Added a timer so that the game won't run as fast as possible - so the
timer actually slows it down quite a bit. Not at all noticeable by human
eyes, of course, except that since it's "slower" its less "blinky" now
that the screen isn't clearing and refreshing at ludacris speeds.