Fixed screen clearing function, added screenshots
Realized when I tried to take screenshots of the game yesterday, the
background was transparent in my terminal - and the game was only
drawing the background when I'd draw in a space. So, the fix was to have
the game draw a black ' ' char on a black background on the entire
screen when calling scr_clear().

Also, added screenshots back in to README.md.