Fixed the screen flickering issue...
By adding a simple redraw flag to the game loop. Not a very elegant
solution, but it works for now. If something changes on the screen, the
flag indicates to the loop that draw needs to be called. At the end of
draw, the flag is then turned false. Simple enough.