Had an idea for the options screen...
I really like the way Goblin Caves handles drawing to the screen (using
a "buffer" - an array of "Glyph" - then drawing that to the screen using
whatever engine). Unfortunately, the way I set it up in GC doesn't work
with the wide (unicode) characters used in Cards. The idea I had is...
why not have two different ways of drawing things? Why not have a
combination of both? Will it work? Switched over to a different branch
to test this crazy idea, and pulled the relevant files over from Goblin
Caves.

Also, I modified the save/load game function to save/load all the settings in
the global settings struct - because at some point the user will
actually be able to set those settings.