Changed "card" linked list structure
Made it into a double linked list - each card now has a next AND a prev
field and it **should** work as intended. Idea is to use the prev field
for checking moves - instead of failing when moving a full sequence,
check each card in sequence to see if that can be moved.