Transferred drawing functions from Goblin Caves
Pulled over the cool, generic, drawing things I wrote for Goblin Caves
to use in this project... which also required me to pull slist.h/c and
rect.h/c from the Toolbox. The fun thing about pulling all the generic
drawing functions is that the result is generic AND reusable enough that
it can go into the toolbox... once I figure out a better name for it.