Win condition check moved...
Played a game, and actually won... then it didn't do the fancy message?
What the heck? Oh right, it was waiting for a keypress to update...

Fixed this. Moved the win condition check to the end of the update
routine so that as soon as the player drops the last card into the right
foundation... the fancy win message is displayed and the buttons are
drawn as inactive.