Added a settings menu and color picker
The settings menu was easy, but the screen buffer drawing needed some
help to draw "transparent" characters... or really just drawing nothing
when the glyph is null. This GREATLY improved using the drawing
functions in term_engine with the screen buffer drawing functions. They
play nicely now.

The color picker is mostly portable, and can probably be thrown in the
toolbox for later. Not sure if I'm happy with the way it looks, but it
works well.

User can now set custom card colors, and the game saves the user's
settings in a local file.