Interface updates, general housekeeping
Cleaned up a bunch of random comments throughout the project, and added
comments in appropriate places.

Added a "pause" menu to Klondike, and a "main" menu to the game - in an
attempt to make the interface more like a real application instead of
something whipped together. The main menu will make more sense when I
add other games, of course.