Changed Klondike struct
So I started work on Penguin - and realized that the Penguin struct was
identical to what I wrote for the Klondike struct... then started
thinking about how I wouldn't change it for really any solitaire game I
could think up. Clearly, that struct needed to be it's own thing.
"Klondike" (defined in klondike.h) was renamed to "Solitaire" (defined
in solitaire.h). Figured this was also a good time to remove all the
"bool" flags I had in the struct and make them proper bit flags.