Began work on Penguin
Added the init, cleanup, and deal functions for Penguin in penguin.c.

Also, started to write a penguin_msg function, realized it was identical
to the klondike_msg function, so wrote a solitaire_msg function that
takes a game of solitaire as the first argument, then the message and
variadics.

In deck.h/c I added a couple functions for returning the rank or suite
flag of a card - basically it just returns the card flags without any of
the suite/rank flags attached.