Penguin draws!
Finished the draw routine for Penguin - screen looks a little too "busy"
though, so I might tweak the layout a couple of times. The important
thing is that I can now see that the penguin_deal() routine is actually
working and putting all the cards where they need to go. Neat.