Penguin main loop, new score saves
Games now keep track of the last score for that game, the high score,
and how many wins the user has for the game. Saving/loading scores
updated in save.c.

Started writing penguin_loop() by copying klondike_loop()... realized
that was silly and a good indicator I should be making that function
generic. Added function pointers to the Solitaire game struct for the
events/update/draw functions for each game, and the game now sets those
during init. The generic solitaire_loop(Solitaire*) is then called by
both penguin_loop() and klondike_loop() with their game pointer. Pretty
cool.

Made a minor modification to the game init functions - they now return a
bool for if the user wants to quit the game entirely when it stops or if
they want to return to the main menu. Not very elegant, but it works
nicely.