Penguin events function added
Also, as is the theme in the last couple commits, I started writing the
penguin_pause function, realized it was exactly the same as the
klondike_pause so... I made it more generic, added it to solitaire, and
all is happy. Created penguin_update.c to hold penguin_update() - which
is the last function needed for the game, and will be created next.