Finished penguin sequence moving rules
Everything works as it should EXCEPT moving a sequence of cards that
would leave the tableau empty. That isn't working. For some reason, it's
not NULL-ing the tableau's card reference.... Oh well it's a start.
Played through a couple games of Penguin even with the little bug - and
its a lot of fun. Scoring needs some tuning. Need to work on the high
score screen - so that it actually shows the high score for Penguin, and
the wins for the other games.