Starting working on a "help" screen...
But I didn't end up doing anything really for all that. Wrote in a
simple setting option to change the design on the back of the cards -
long overdue. Also, realized this whole time the back of the cards would
have looked a LOT better if I switched the fg/bg colors.

The help screen does exist, but mostly as a placeholder right now. I'll
add some text and stuff to it... eventually.