Penguin scoring adjustments
I think I figured out a way to make the scoring a bit better for
Penguin. The player gets 5 points if they make a valid move from one
tableau card to another tableau card - but not if they make the move
from the cells... what this does is incentivise building on the
tableaus, and makes it possible to play full games with different
finishing scores.