Penguin QoL updates
After playing through numerous games of penguin, a couple of things were
bothering me - so I fixed them. The user can now move a card from a cell
to another cell, swapping places with whatever card is there. Super easy
update, and boy does it make things nicer.

Additionally, when the user moves cards to the foundation it checks to
see if the card is in sequence - if it is it moves ALL the cards in
sequence to the foundation, and gives the user 15 points per card.