cribbage_draw() "finished"
At least mostly. It looks pretty good! The pegs are drawn in the
appropriate places when score is increased for both computer and player.
Need to add some more options in for changing the color of the
pegs/cribbage board.

Started working on the "discard" state as well - prompting the player to
select two cards to add to the crib.