Play and show states roughed in for Cribbage
None of the logic is there (yet) but the players alternate putting cards
on the table, and when they are out of cards the state changes to the
"show" state which displays a message with the points for the hands
using a CribScore object (defined in cribscore.c).

I originally wrote the CribScore stuff in 2022, and some of it is silly
brute force and other parts are PFM and I wish I wrote comments
explaining how I figured out what I did. Some of it needed revamping to
play nicely, and I've been keeping track of what I've changed from the
OG project.