Split the massive cribbage.c file...
Into draw/events/update files, as is tradition.

Also took the time to explain to myself some code I wrote in CribScore
in 2022 that was absolutely PFM and I have no idea where I got the idea
from. In doing so I found and (possibly) fixed the bug in counting runs.