Minor changes to Cribbage win, term engine update
Added a "blinky" cursor char getting function to term engine - basically
while the computer blocks waiting for a char input,  it draws a single █
character with the "blink" style at a specified x,y position (ideally
right after or right at where the prompt for the character input is
printed). Useful!

Made the cribbage "win" condition it's own state, so that a happy "You
won!" banner can be easily drawn at the top of the screen when the
player wins (nothing is drawn when they lose). Should also make it easy
to track computer/player wins if I want to add a counter on/next to the
board.