Minor drawing fixes, cribbage win fixes
"Drawing" transparent characters to the screen buffer and then drawing
on top of it works well EXCEPT for when a terminal has the background
set with transparency. Easy fix is to "separate" the drawing - draw the
screen buffer stuff first THEN draw the other things on the terminal.

Forgot to reset a couple of things when starting a new round of
cribbage, and needed to move cards from the board back to the stock.