Fixed bug in Cribbage
The computer cheats for some reason. If it has a card that matches whats
on the table, it will play that card regardless if it goes over 31.
Realized this is an easy fix... just add a check after it chooses a card
to make sure that the card it chose plus whats on the board doesn't
exceed 31 - if it does, slap a random card on the table.