Penguin improvements
I made the next card in rank from the tableaus for penguin "blink",
which I think makes it easier/quicker to spot. I will probably add a
toggle key to turn this on and off. Also - I am constantly firing this
up on my terminal and playing it, figured this was a nice way to dive
back into the project/continue work on it.